+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.5.0] - 02-25-25
 
+### [2.5.0-beta.15]
+- Added youtube fallback options to Video Manager
+- Added LocalOffset property to Sync Video Player API
+
 ### [2.5.0-beta.14]
 - Fixed lock mute option in Audio Channel script
 - Added 'Use Reserved Audio Source' option for AudioLink to audio profiles, overriding selected channel